Working MNIST predicting model, some fixes & add cuda pckg

+++ init.sh
@@ -1,13 +1,25 @@
 #!/bin/bash
 
 # Create a virtual environment
-python3 -m venv .venv
+if [ -d ".venv" ]; then
+	echo "Virtual environment already exists..."
+	continue # Skip the virtual environment creation
+else
+	python3 -m venv .venv
+fi
 
-# Activate the virtual environment
-source .venv/bin/activate
+if [ -f ".venv/activate.lock" ]; then
+	echo "Virtual enviroment already running..."
+	continue # Skip the virtual environment activation
+else
+	# Create a lock file
+	touch .venv/activate.lock
+	# Activate the virtual environment
+	source .venv/bin/activate
+	# Upgrade pip
+	pip install --upgrade pip
 
-# Upgrade pip
-pip install --upgrade pip
+	# Install the required packages
+	pip install -r requirements.txt
+fi
 
-# Install the required packages
-pip install -r requirements.txt
